# api/openapi.yaml
openapi: 3.0.3
info:
  title: FitCore API
  version: 0.1.0
  description: >-
    Gym management platform covering branches, members, memberships, group
    classes, bookings, attendance check-in/out, invoicing, staff and trainers.
    Contract-first: routes are documented ahead of their HTTP adapters, which
    are implemented incrementally. Every operation except health, metrics and
    token issuance requires a short-lived Bearer access token (JWT) carrying
    the caller's permission set; refresh tokens rotate via /auth/refresh.
security:
  - BearerAuth: []
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /members/{memberId}/memberships:
    parameters:
      - $ref: "#/components/parameters/MemberID"
    get:
      tags: [memberships]
      summary: List a member's membership history
      operationId: listMemberMemberships
      x-permission: memberships:read
      responses:
        "200":
          description: List of memberships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Membership"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /memberships:
    post:
      tags: [memberships]
      summary: Purchase a membership
      description: Creates a membership for a member from a package, starting now.
      operationId: createMembership
      x-permission: memberships:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipCreateRequest"
      responses:
        "201":
          description: Membership created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /memberships/{membershipId}:
    parameters:
      - $ref: "#/components/parameters/MembershipID"
    get:
      tags: [memberships]
      summary: Get a membership by ID
      operationId: getMembership
      x-permission: memberships:read
      responses:
        "200":
          description: Membership found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [memberships]
      summary: Update a membership
      description: Used to freeze, resume or expire a membership.
      operationId: updateMembership
      x-permission: memberships:update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdateRequest"
      responses:
        "200":
          description: Membership updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /packages:
    get:
      tags: [packages]
      summary: List membership packages
      operationId: listPackages
      x-permission: packages:read
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [packages]
      summary: Create a membership package
      operationId: createPackage
      x-permission: packages:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageCreateRequest"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /packages/{packageId}:
    parameters:
      - $ref: "#/components/parameters/PackageID"
    get:
      tags: [packages]
      summary: Get a package by ID
      operationId: getPackage
      x-permission: packages:read
      responses:
        "200":
          description: Package found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [packages]
      summary: Update a package
      operationId: updatePackage
      x-permission: packages:update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdateRequest"
      responses:
        "200":
          description: Package updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /classes:
    get:
      tags: [classes]
      summary: List classes
      description: Optionally filter by branch and trainer.
      operationId: listClasses
      x-permission: classes:read
      parameters:
        - $ref: "#/components/parameters/BranchIDQuery"
        - $ref: "#/components/parameters/TrainerIDQuery"
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [classes]
      summary: Create a class
      operationId: createClass
      x-permission: classes:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassCreateRequest"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /classes/{classId}:
    parameters:
      - $ref: "#/components/parameters/ClassID"
    get:
      tags: [classes]
      summary: Get a class by ID
      operationId: getClass
      x-permission: classes:read
      responses:
        "200":
          description: Class found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [classes]
      summary: Update a class
      operationId: updateClass
      x-permission: classes:update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdateRequest"
      responses:
        "200":
          description: Class updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [classes]
      summary: Delete a class
      operationId: deleteClass
      x-permission: classes:delete
      responses:
        "204":
          description: Class deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Short-lived access token issued by POST /auth/login, signed with
        HMAC-SHA256. Claims include sub (staff id), jti, perms (permission
        set) and exp.
  parameters:
...
    MemberID:
      name: memberId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    MembershipID:
      name: membershipId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    PackageID:
      name: packageId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    ClassID:
      name: classId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
    BranchIDQuery:
      name: branchId
      in: query
      required: false
      description: Filter by branch
      schema:
        $ref: "#/components/schemas/UUID"
    TrainerIDQuery:
      name: trainerId
      in: query
      required: false
      description: Filter by trainer
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
    BadRequest:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    UUID:
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    Timestamp:
      type: string
      format: date-time
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message, safe for clients.
        code:
          type: string
          nullable: true
          description: Stable machine-readable error code.
      example:
        error: "member not found"
        code: "member_not_found"
...
    MembershipStatus:
      type: string
      enum: [active, frozen, expired]
    Membership:
      type: object
      required:
        - id
        - memberId
        - packageId
        - branchId
        - startsAt
        - expiresAt
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        memberId:
          $ref: "#/components/schemas/UUID"
        packageId:
          $ref: "#/components/schemas/UUID"
        branchId:
          $ref: "#/components/schemas/UUID"
        startsAt:
          $ref: "#/components/schemas/Timestamp"
        expiresAt:
          $ref: "#/components/schemas/Timestamp"
        status:
          $ref: "#/components/schemas/MembershipStatus"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    MembershipCreateRequest:
      type: object
      required: [memberId, packageId, branchId]
      properties:
        memberId:
          $ref: "#/components/schemas/UUID"
        packageId:
          $ref: "#/components/schemas/UUID"
        branchId:
          $ref: "#/components/schemas/UUID"
    MembershipUpdateRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/MembershipStatus"
        expiresAt:
          $ref: "#/components/schemas/Timestamp"
    Package:
      type: object
      required: [id, name, durationDays, priceCents, currency, active, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        durationDays:
          type: integer
          minimum: 1
        priceCents:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: BHD
        active:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    PackageCreateRequest:
      type: object
      required: [name, durationDays, priceCents, currency]
      properties:
        name:
          type: string
        durationDays:
          type: integer
          minimum: 1
        priceCents:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
    PackageUpdateRequest:
      type: object
      properties:
        name:
          type: string
        durationDays:
          type: integer
          minimum: 1
        priceCents:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        active:
          type: boolean
    Class:
      type: object
      required: [id, branchId, name, startsAt, endsAt, capacity, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        branchId:
          $ref: "#/components/schemas/UUID"
        trainerId:
          allOf:
            - $ref: "#/components/schemas/UUID"
          nullable: true
        name:
          type: string
        startsAt:
          $ref: "#/components/schemas/Timestamp"
        endsAt:
          $ref: "#/components/schemas/Timestamp"
        capacity:
          type: integer
          minimum: 1
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    ClassCreateRequest:
      type: object
      required: [branchId, name, startsAt, endsAt, capacity]
      properties:
        branchId:
          $ref: "#/components/schemas/UUID"
        trainerId:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        startsAt:
          $ref: "#/components/schemas/Timestamp"
        endsAt:
          $ref: "#/components/schemas/Timestamp"
        capacity:
          type: integer
          minimum: 1
    ClassUpdateRequest:
      type: object
      properties:
        name:
          type: string
        startsAt:
          $ref: "#/components/schemas/Timestamp"
        endsAt:
          $ref: "#/components/schemas/Timestamp"
        capacity:
          type: integer
          minimum: 1
...